+++ primitives/rpc/src/lib.rs
@@ -1,12 +1,10 @@
 #![cfg_attr(not(feature = "std"), no_std)]
 
-pub mod api;
 pub mod chain;
 pub mod identity;
 pub mod transactions;
 pub mod verifier;
 
-pub use api::*;
 pub use chain::*;
 pub use identity::*;
 pub use transactions::*;
